+++ src/parsing/clojure/parser.rs
@@ -106,9 +106,9 @@ impl ClojureParser {
     fn range_from_node(node: &Node) -> Range {
         Range::new(
             node.start_position().row as u32,
-            node.start_position().column as u16,
+            node.start_position().column as u32,
             node.end_position().row as u32,
-            node.end_position().column as u16,
+            node.end_position().column as u32,
         )
     }
